fix bugs in the admin ms

+++ src/modules/discount/controller/deleteDiscount.ts
@@ -1,4 +1,5 @@
 import { AdminDiscount } from '../../../models/discount';
+import { Product } from '../../../models/product';
 import { Request, Response } from 'express'
 
 export const deleteDiscountOnProduct = async (req: Request, res: Response) => {
@@ -12,6 +13,12 @@ export const deleteDiscountOnProduct = async (req: Request, res: Response) => {
       if (!discount) {
          return res.json({ message: "discount not found" })
       }
+      const product = await Product.findOne({ _id: discount._product })
+      if (!product) {
+         return res.json({ message: 'discount on product not found' })
+      }
+      product.discount = undefined
+      await product.save()
       await AdminDiscount.findByIdAndDelete({ _id: _discountId });
       return res.status(200).json({ success: true, message: "Discount deleted successfully" })
    } catch (error) {
